Fixed issue #2 (Python 3.x compatibility).

+++ pywinauto/controls/HwndWrapper.py
@@ -1689,9 +1689,9 @@ def _perform_click_input(
 
     # set the default coordinates
     if coords[0] is None:
-        coords[0] = ctrl.Rectangle().width() / 2
+        coords[0] = int(ctrl.Rectangle().width() / 2)
     if coords[1] is None:
-        coords[1] = ctrl.Rectangle().height() / 2
+        coords[1] = int(ctrl.Rectangle().height() / 2)
 
     if not absolute:
         screen_coords = win32structures.POINT()
